Update v.6b
An end user requested that the script also be able to check for participants who entered shelter within three months of being placed into permanent housing.  Since this was only a minor change from the check_for_entry_3_months_prior_to_placement method a new method called check_for_placement_3_months_prior_to_shelter_entry was created providing this new functionality, the process method was split into two separate methods to accommodate either option.  When running the script now, the if __name__ == "__main__" call will need to be adjusted to make sure that the correct process method is called.

Additionally testing was conducted on the check_for_entry_3_months prior_to_placement version of the process method and and all the data looked exactly as expected.

Provided no new feature options are selected then after testing the new methods output this will move to version 1.0.  If it seems like this is something that will need to be run frequently I will make a gui for selecting between the two processing methods, though this will be done as a GUI fork.